+++ src/main/java/com/echo/echo/domain/space/error/SpaceSuccessCode.java
@@ -1,4 +1,4 @@
-package com.echo.echo.common.exception.codes.error;
+package com.echo.echo.domain.space.error;
 
 import com.echo.echo.common.exception.BaseCode;
 import com.echo.echo.common.exception.CommonReason;
@@ -7,16 +7,14 @@ import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
 /**
- * CommonErrorCode 열거형(enum)은 공통적으로 사용할 에러 코드를 정의
+ * SpaceSuccessCode 열거형(enum)은 스페이스 관련 성공 메시지를 정의
  */
+
 @Getter
 @AllArgsConstructor
-public enum CommonErrorCode implements BaseCode {
+public enum SpaceSuccessCode implements BaseCode {
+    SPACE_DELETE(3, "스페이스 삭제 완료입니다.", "스페이스 삭제가 성공적으로 완료되었습니다.");
 
-    FAIL(HttpStatus.INTERNAL_SERVER_ERROR, -1, "실패했습니다.", "서버 에러"),
-    UNAUTHORIZED(HttpStatus.UNAUTHORIZED, 401, "로그인이 안되어있거나 만료된 사용자입니다.", "비정상적인 토큰입니다.");
-
-    private final HttpStatus status;
     private final int code;
     private final String msg;
     private final String remark;
@@ -24,7 +22,7 @@ public enum CommonErrorCode implements BaseCode {
     @Override
     public CommonReason getCommonReason() {
         return CommonReason.builder()
-            .status(status)
+            .status(HttpStatus.OK)
             .code(code)
             .msg(msg)
             .build();
